requirements.txt
Django==3.2.10
psycopg2-binary==2.8.5
Pillow==7.1
djangorestframework==3.11
django-cors-headers==3.1.1
django-rest-auth==0.9.5
django-allauth==0.41.0
